requirements.txt
# --- Core ML & Data Science ---
torch~=2.5.1
torchvision~=0.20.1
torchaudio~=2.5.1
timm~=1.0.19
numpy~=2.2.5
pandas~=2.3.1
scikit-learn~=1.7.1
Pillow~=11.3.0
tqdm~=4.67.1

# --- Application & API Frameworks ---
streamlit~=1.48.1
altair~=5.5.0
fastapi~=0.116.1
uvicorn[standard]~=0.35.0
python-multipart~=0.0.20

# --- Configuration ---
PyYAML~=6.0.2